refactor: move zizmor check to a new Security family and page
Begin implementing the Security page (#808). The zizmor check moves out
of the GitHub Actions family into a dedicated Security family (GH106 ->
SEC001), and the "Linting your workflows" guide section moves into a new
docs/guides/security.md page, which is intended to grow with more
security topics (pre-commit hash pinning, lock files, cooldowns,
pip/uv audit, SBOMs).
