+++ docs/myst.yml
@@ -30,6 +30,7 @@ project:
         - file: guides/gha_basic.md
         - file: guides/gha_pure.md
         - file: guides/gha_wheels.md
+        - file: guides/security.md
         - file: guides/tasks.md
     - file: principles/index.md
       children:
